+++ src/plugins/data_source_management/public/components/data_source_selector/data_source_selector.tsx
@@ -38,6 +38,7 @@ interface DataSourceSelectorState {
 export interface DataSourceOption {
   label: string;
   id: string;
+  checked?: string;
 }
 
 export class DataSourceSelector extends React.Component<